[#18] Cleaned up routes file, removed dead code
The routes file is now easier to read. The following have been removed:

* Asynchronously offer search suggestions

An ajax call is being made to spell check after the search results are
loaded, but the result isn't appended correctly to the suggestion.  We
calculate the spelling suggestion server-side anyway, so this has been
removed.

* Unused ArticlesController#missing action

Leftover from when missing articles were not handled correctly.