[#18] Remove dead (unused) code
The following have been removed:

* ApplicationController#stop_words

Leftover from the attempt to asynchronously offer search suggestions

* CategoriesController#missing

Explicit redirect in #show obviates this

* ApplicationHelper #categories_list, #catgeory_by_name

Originally used in the categories sidebar, now unused.

* ArticlesHelper #article_list_tag, #missing_article_list_tag

Only used in the now deleted articles list view.

* Article#with_category scope

Potentially never used

* Article#find_by_friendly_id

Used to be used in the fake popular articles on the homepage. Now we
have real popular articles so this is gone.

* Article#allContent

Unused, and awful.

* Article#publish

Added but never used?

* Article#content_to_markdown

Was used at some point to auto-convert the HTML to markdown, is no
longer called anywhere.

* Article#content_md_to_html

BlueCloth hasn't been used for a long time.

* Article#hits, Category#hits

Added for console convenience, never used in the app

* Article#analyse_all

Used to be used in a rake task, now unused.

* Keyword#create_all

Only used in a test

* articles/index.html.erb.bak

Backup files redundant when using version control.