[#37] Fixed button styles when editing article as writer or editor
Necessitated changing the javascript a bit so we can use the normal
ActiveAdmin classes on the buttons without it triggering the usual form
submission.

Removed the confusing "Commit to editing" input for editors and replaced
it with a dropdown of users: editors can now freely assign writers to
articles.

Removed the also confusing and broken "Review by <email>" status field
for admin users, so that now you cannot assign an editor to an article.