[#47] Fix preview with versioning: QuickAnswers#preview for previews.
Adding versioning broke previews.

Implementation of preview was: if the article is a draft and you are
able to view drafts, you're seeing a preview.

This approach no longer works since the most recent version is not
necessarily the one we want to #show.

The solution is to add an explicit controller action for previews. In
doing so, QuickAnswersController has been refactored to reduce
duplication and keep the actions small.

CAVEATS:

Currently only working with QuickAnswers.